Revert "Convert to classes"
