+++ test.py
@@ -1,0 +1,12 @@
+# purely made to test out LEGB rule in python
+
+def sum_numbers(a, b):
+    return a + b
+
+
+def print_sum(a, b):
+    print(sum_numbers(a, b))
+
+
+print_sum(1, 3)
+# 4
